S-44: finalize refactor of moving logic into controller files
